% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{add_question_from_bank}
\alias{add_question_from_bank}
\title{Add questions from a question bank to a quiz}
\usage{
add_question_from_bank(
  quiz_id,
  bank_id,
  title = NULL,
  n = 1L,
  points = 1L,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{quiz_id}{The id of the quiz to add questions to.}

\item{bank_id}{The id of the question bank to pull questions from.}

\item{title}{(Optional) The title of the quiz group to create for these questions. If \code{NULL} (default), no quiz group will be created and questions will be added to the quiz without a group.}

\item{n}{The number of questions to randomly select from the question bank. Defaults to \code{1}.}

\item{points}{The number of points each question added from the bank should be worth. Defaults to \code{1}.}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\value{
The API response from adding the questions.
}
\description{
Add questions from a question bank to a quiz
}
