% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer.R
\name{answer_essay}
\alias{answer_essay}
\title{Answer for an essay question}
\usage{
answer_essay()
}
\description{
Answer for an essay question
}
\seealso{
Other answer-functions: 
\code{\link{answer_matching}()},
\code{\link{answer_mcq}()},
\code{\link{answer_multiple}()},
\code{\link{answer_none}()},
\code{\link{answer_num}()},
\code{\link{answer_single}()},
\code{\link{answer_text}()},
\code{\link{answer_true_false}()},
\code{\link{answer_upload_file}()}
}
\concept{answer-functions}
