% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer.R
\name{answer_matching}
\alias{answer_matching}
\title{Answer for a matching question}
\usage{
answer_matching(left, right, extra_choices = "")
}
\arguments{
\item{left}{Character vector of the left-hand side items to be matched.}

\item{right}{Character vector of the right-hand side items to be matched. Must be the same length as \code{left}.}

\item{extra_choices}{Character vector of extra choices that can be used as incorrect matches. Optional.}
}
\description{
Answer for a matching question
}
\seealso{
Other answer-functions: 
\code{\link{answer_essay}()},
\code{\link{answer_mcq}()},
\code{\link{answer_multiple}()},
\code{\link{answer_none}()},
\code{\link{answer_num}()},
\code{\link{answer_single}()},
\code{\link{answer_text}()},
\code{\link{answer_true_false}()},
\code{\link{answer_upload_file}()}
}
\concept{answer-functions}
