% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{answer_mcq}
\alias{answer_mcq}
\title{Answer for a multiple choice question}
\usage{
answer_mcq(choices, correct, multiple = length(correct) > 1)
}
\arguments{
\item{choices}{Character vector of answer choices.}

\item{correct}{Character vector of correct answer(s). Should be one element for single-answer multiple choice questions and can be multiple elements for multiple-answer multiple choice questions.}

\item{multiple}{Logical. Whether the question allows multiple correct answers. Defaults to \code{TRUE} if \code{correct} has more than one element.}
}
\description{
Answer for a multiple choice question
}
\seealso{
Other answer-functions: 
\code{\link{answer_essay}()},
\code{\link{answer_matching}()},
\code{\link{answer_multiple}()},
\code{\link{answer_none}()},
\code{\link{answer_num}()},
\code{\link{answer_text}()},
\code{\link{answer_true_false}()},
\code{\link{answer_upload_file}()}
}
\concept{answer-functions}
