% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer.R
\name{answer_num}
\alias{answer_num}
\alias{answer_num_precision}
\alias{answer_num_range}
\title{A numerical answer question with an exact answer}
\usage{
answer_num(value, tol = 0)

answer_num_precision(value, precision = 0L)

answer_num_range(lower, upper = lower)
}
\arguments{
\item{value}{The correct numerical answer.}

\item{tol}{The acceptable error margin for the answer. Defaults to 0 for an exact answer.}

\item{precision}{The number of decimal places the answer must be correct to.}

\item{lower}{The lower bound of the acceptable answer range.}

\item{upper}{The upper bound of the acceptable answer range.}
}
\description{
A numerical answer question with an exact answer
}
\seealso{
Other answer-functions: 
\code{\link{answer_essay}()},
\code{\link{answer_matching}()},
\code{\link{answer_mcq}()},
\code{\link{answer_multiple}()},
\code{\link{answer_none}()},
\code{\link{answer_single}()},
\code{\link{answer_text}()},
\code{\link{answer_true_false}()},
\code{\link{answer_upload_file}()}
}
\concept{answer-functions}
