% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer.R
\name{answer_single}
\alias{answer_single}
\title{Create an answer object for a quiz question}
\usage{
answer_single(x, y = NULL, tol = 0, precision = 0)
}
\arguments{
\item{x}{The correct answer(s) for the question. See details for the expected format based on the type of question.}

\item{y}{Additional parameter for certain question types. See details for the expected format based on the type of question.}

\item{tol}{The acceptable error margin for the answer. Defaults to 0 for an exact answer.}

\item{precision}{The number of decimal places the answer must be correct to.}
}
\description{
This function is a wrapper that creates the appropriate answer object based on the input. It can handle multiple choice questions, true/false questions, short answer questions, numerical questions (with exact answers, precision-based answers, or range-based answers), and matching questions. The function will automatically determine the type of question based on the format of the input and create the corresponding answer object using the appropriate helper functions.
}
\details{
The function uses the following logic to determine the type of question:
\itemize{
\item If \code{x} is a logical value and \code{y} is NULL, it creates a true/false question using \code{answer_true_false()}.
\item If \code{x} is a single numeric value and \code{y} is NULL, it creates a numerical question with an exact answer using \code{answer_num()}. If \code{tol} is greater than 0, it creates a numerical question with an error margin. If \code{precision} is greater than 0, it creates a numerical question with a specified precision.
\item If \code{x} is a numeric vector of length 2, it creates a numerical question with a range-based answer using \code{answer_num_range()}. Alternatively, if \code{x} is a single numeric value and \code{y} is a single numeric value, it also creates a numerical question with a range-based answer using \code{answer_num_range()}, where \code{x} is the lower bound and \code{y} is the upper bound.
\item If \code{x} is a named character vector, it creates a matching question using \code{answer_matching()}. Any extra choices for the matching question can be provided in \code{y}.
\item If \code{x} and \code{y} are character vectors, it creates a multiple choice question using \code{answer_mcq()}.
\item If the input format does not match any of the above cases, the function throws an error indicating that the answer format is unsupported.
}
}
\seealso{
Other answer-functions: 
\code{\link{answer_essay}()},
\code{\link{answer_matching}()},
\code{\link{answer_mcq}()},
\code{\link{answer_multiple}()},
\code{\link{answer_none}()},
\code{\link{answer_num}()},
\code{\link{answer_text}()},
\code{\link{answer_true_false}()},
\code{\link{answer_upload_file}()}
}
\concept{answer-functions}
