% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer.R
\name{answer_text}
\alias{answer_text}
\title{A short answer question}
\usage{
answer_text(correct)
}
\arguments{
\item{correct}{The correct answer text.}
}
\description{
A short answer question
}
\seealso{
Other answer-functions: 
\code{\link{answer_essay}()},
\code{\link{answer_matching}()},
\code{\link{answer_mcq}()},
\code{\link{answer_multiple}()},
\code{\link{answer_none}()},
\code{\link{answer_num}()},
\code{\link{answer_single}()},
\code{\link{answer_true_false}()},
\code{\link{answer_upload_file}()}
}
\concept{answer-functions}
