% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{answer_true_false}
\alias{answer_true_false}
\title{True or False Question}
\usage{
answer_true_false(correct = TRUE)
}
\arguments{
\item{correct}{Logical. Whether the correct answer is \code{TRUE} or \code{FALSE}.}
}
\description{
True or False Question
}
\seealso{
Other answer-functions: 
\code{\link{answer_essay}()},
\code{\link{answer_matching}()},
\code{\link{answer_mcq}()},
\code{\link{answer_multiple}()},
\code{\link{answer_none}()},
\code{\link{answer_num}()},
\code{\link{answer_text}()},
\code{\link{answer_upload_file}()}
}
\concept{answer-functions}
