% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer.R
\name{answer_upload_file}
\alias{answer_upload_file}
\title{A file upload question}
\usage{
answer_upload_file()
}
\description{
A file upload question
}
\seealso{
Other answer-functions: 
\code{\link{answer_essay}()},
\code{\link{answer_matching}()},
\code{\link{answer_mcq}()},
\code{\link{answer_multiple}()},
\code{\link{answer_none}()},
\code{\link{answer_num}()},
\code{\link{answer_single}()},
\code{\link{answer_text}()},
\code{\link{answer_true_false}()}
}
\concept{answer-functions}
