% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{attempt-options}
\alias{attempt-options}
\alias{opt_show}
\alias{opt_hide}
\alias{opt_shuffle}
\title{Options for showing quiz results or answers}
\usage{
opt_show(last_attempt = FALSE, one_time = FALSE, at = NULL)

opt_hide(at = NULL)

opt_shuffle()
}
\arguments{
\item{last_attempt}{logical. Whether to show results/answers only after the last attempt is submitted. Only valid if the number of attempts is greater than 1. Defaults to \code{FALSE}.}

\item{one_time}{logical. Whether to show results/answers only immediately after submission.}

\item{at}{POSIXct. Date and time to control when results/answers become visible or hidden. If \code{show = TRUE}, results/answers become visible at this date and time. If \code{hide = TRUE}, results/answers become hidden at this date/time.}
}
\value{
A list of options to be passed to the \code{results} or \code{answer} parameters of \code{\link[=attempt_options]{attempt_options()}}.
}
\description{
Options for showing quiz results or answers
}
