% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{attempt_options}
\alias{attempt_options}
\title{A quiz attempt options}
\usage{
attempt_options(
  n = 1,
  time = NULL,
  due = NULL,
  lock = NULL,
  unlock = NULL,
  results = opt_show(at = due),
  answer = opt_show(at = due),
  score = c("highest", "latest"),
  access_code = NULL,
  ip_filter = NULL
)
}
\arguments{
\item{n}{The number of allowed attempts. Set to \code{-1} or \code{Inf} for unlimited. Defaults to \code{1}.}

\item{time}{Time limit for each attempt in minutes. Set to \code{NULL} for no limit (default).}

\item{due}{POSIXct. Due date and time for the quiz.}

\item{lock}{POSIXct. Lock date and time for the quiz.}

\item{unlock}{POSIXct. Unlock date and time for the quiz.}

\item{results}{A list of options controlling when quiz results are visible to students. See \code{\link[=opt_show]{opt_show()}} and \code{\link[=opt_hide]{opt_hide()}}.}

\item{answer}{A list of options controlling when correct answers are visible to students. See \code{\link[=opt_show]{opt_show()}}, \code{\link[=opt_hide]{opt_hide()}} and \code{\link[=opt_shuffle]{opt_shuffle()}}.}

\item{score}{Scoring policy for multiple attempts. Allowed values: "\code{highest}" (default), "\code{latest}".}

\item{access_code}{Password required to access the quiz. Set to \code{NULL} for no restriction (default).}

\item{ip_filter}{Restrict access to computers in a specified IP range. Filters can be a comma-separated list of addresses, or address/mask. Examples: \code{"192.168.1.1,192.168.1.2"} or \code{"192.168.1.0/24"}. Set to \code{NULL} for no restriction (default).}
}
\value{
A list of options to be passed to \code{\link[=create_quiz]{create_quiz()}}.
}
\description{
A quiz attempt options
}
