% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas}
\alias{canvas}
\alias{set_canvas_course_id}
\alias{get_canvas_course_id}
\alias{set_canvas_url}
\alias{get_canvas_url}
\alias{set_canvas_token}
\alias{get_canvas_token}
\title{Set and get Canvas course, URL, and token}
\usage{
set_canvas_course_id(course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"))

get_canvas_course_id()

set_canvas_url(url = Sys.getenv("CANVASQUIZ_URL"))

get_canvas_url()

set_canvas_token(token = Sys.getenv("CANVASQUIZ_TOKEN"))

get_canvas_token()
}
\arguments{
\item{course_id}{The Canvas course ID to use for API calls.}

\item{url}{The base URL for the Canvas instance (e.g., "https://
canvas.instructure.com").}

\item{token}{The API token for authenticating with the Canvas API.}
}
\description{
These functions set and retrieve the Canvas course ID, URL, and token from
environment variables. This allows for easy configuration of API calls to the
Canvas API without hardcoding sensitive information in the code.
}
