% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{create_question}
\alias{create_question}
\title{Create a Quiz Question}
\usage{
create_question(
  text = NULL,
  answers = NULL,
  points = 1,
  quiz_id = last_quiz_id(),
  position = NULL,
  title = NULL,
  correct_comments = NULL,
  incorrect_comments = NULL,
  neutral_comments = NULL,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{text}{Character. The text of the question.}

\item{answers}{A list of answer objects.}

\item{points}{The maximum amount of points received for answering this question correctly.}

\item{quiz_id}{The id of the quiz to add the question to.}

\item{position}{An integer specifying the order in which the question will be displayed in the quiz. Doesn't seem to work.}

\item{title}{The name of the question.}

\item{correct_comments}{Comment to display if the student answers the question correctly.}

\item{incorrect_comments}{Comment to display if the student answers incorrectly.}

\item{neutral_comments}{Comment to display regardless of how the student answered.}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\value{
A list representing the quiz question.
}
\description{
This function creates a quiz question directly into Canvas. You can create a variety of question types, including multiple choice, short answer, numerical answer, matching, essay, file upload, and text-only questions.
}
\details{
The question comments is correctly entered, however, Canvas doesn't seem to display them initially. Manually editing the quiz and saving it again seems to fix the issue and then the comments are displayed as expected.
}
