% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{create_quiz}
\alias{create_quiz}
\title{Create a Quiz}
\usage{
create_quiz(
  title,
  description = NULL,
  quiz_type = c("practice_quiz", "assignment", "graded_survey", "survey"),
  one_question_at_a_time = c("no", "yes", "yes_but_cant_go_back"),
  attempts = attempt_options(),
  publish = FALSE,
  visible = c("everyone", "none"),
  assignment_group_id = NULL,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{title}{The quiz title (required).}

\item{description}{A description of the quiz.}

\item{quiz_type}{The type of quiz. Allowed values: "\code{practice_quiz}", "\code{assignment}", "\code{graded_survey}", "\code{survey}".}

\item{one_question_at_a_time}{Whether to show one question at a time. Allowed values: "\code{no}" (default), "\code{yes}", "\code{yes_but_cant_go_back}".}

\item{attempts}{A list of options controlling quiz attempts, as created by the \code{\link[=attempt_options]{attempt_options()}} function.}

\item{publish}{Whether to publish the quiz immediately. Defaults to \code{FALSE}.}

\item{visible}{Who this quiz is visible to. Allowed values: "\code{everyone}" (default) or "\code{none}".}

\item{assignment_group_id}{integer. The assignment group ID to put the assignment in. Defaults to the top assignment group in the course. Only valid if \code{quiz_type} is "\code{assignment}" or "\code{graded_survey}".}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\value{
The quiz id.
}
\description{
Create a quiz with a number of customizable parameters.
}
