% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{delete_all_quizzes}
\alias{delete_all_quizzes}
\title{Delete all quizzes}
\usage{
delete_all_quizzes(
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\description{
Use with caution! This function will delete all quizzes in the course after asking for confirmation. It will not delete quiz question banks.
}
