% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{delete_quiz_question}
\alias{delete_quiz_question}
\title{Delete a quiz question}
\usage{
delete_quiz_question(
  quiz_id,
  question_id,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{quiz_id}{The id of the quiz to retrieve questions from.}

\item{question_id}{The id of the question to delete.}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\description{
Delete a quiz question
}
