% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last-ids}
\alias{last-ids}
\alias{last_quiz_id}
\alias{last_question_id}
\title{Get the ID of the most recently created quiz in a course}
\usage{
last_quiz_id()

last_question_id()
}
\value{
The ID of the most recently created quiz in the specified course.
}
\description{
Get the ID of the most recently created quiz in a course
}
