% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{list_quizzes}
\alias{list_quizzes}
\title{List quizzes in a course}
\usage{
list_quizzes(
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\value{
A data frame of quizzes with their details.
}
\description{
Currently only a maximum of 100 quizzes will be returned.
}
