% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{quiz_questions}
\alias{quiz_questions}
\title{Get quiz questions}
\usage{
quiz_questions(
  quiz_id,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{quiz_id}{The id of the quiz to retrieve questions from.}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\value{
A list of quiz questions.
}
\description{
Get quiz questions
}
