% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{quiz_statistics}
\alias{quiz_statistics}
\title{Get quiz statistics}
\usage{
quiz_statistics(
  quiz_id,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{quiz_id}{The ID of the quiz to retrieve statistics for.}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\description{
Get quiz statistics
}
