% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{regrade}
\alias{regrade}
\title{Regrade quiz submissions based on answer}
\usage{
regrade(
  question_id,
  quiz_id,
  answer = NULL,
  tol = 0,
  fudge_points = NULL,
  score = NULL,
  comment = NULL,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{question_id}{The ID of the question to regrade.}

\item{answer}{The answer to regrade. Can be numeric or character.}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\description{
Regrade quiz submissions based on answer
}
