% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission.R
\name{submission_update}
\alias{submission_update}
\title{Update a quiz submission}
\usage{
submission_update(
  submission_id,
  fudge_points = NULL,
  question_id = NULL,
  score = NULL,
  comment = NULL,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{submission_id}{The ID of the quiz submission to retrieve details for.}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\description{
Update a quiz submission
}
