% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{update_question}
\alias{update_question}
\title{Update a quiz question}
\usage{
update_question(
  question_id,
  quiz_id,
  text = NULL,
  answers = NULL,
  points = NULL,
  position = NULL,
  title = NULL,
  correct_comments = NULL,
  incorrect_comments = NULL,
  neutral_comments = NULL,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{quiz_id}{The id of the quiz to add the question to.}

\item{text}{Character. The text of the question.}

\item{answers}{A list of answer objects.}

\item{points}{The maximum amount of points received for answering this question correctly.}

\item{position}{An integer specifying the order in which the question will be displayed in the quiz. Doesn't seem to work.}

\item{title}{The name of the question.}

\item{correct_comments}{Comment to display if the student answers the question correctly.}

\item{incorrect_comments}{Comment to display if the student answers incorrectly.}

\item{neutral_comments}{Comment to display regardless of how the student answered.}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\description{
Update a quiz question
}
