% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{upload_file}
\alias{upload_file}
\title{Upload a file to a folder}
\usage{
upload_file(
  file,
  folder_id,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{file}{The path to the file to upload.}

\item{folder_id}{The ID of the folder to upload the file to.}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\description{
Upload a file to a folder
}
