% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{answer_fill_in_multiple_blanks}
\alias{answer_fill_in_multiple_blanks}
\title{Answer for a fill-in-multiple-blanks question}
\usage{
answer_fill_in_multiple_blanks(answers, blank_ids)
}
\description{
Answer for a fill-in-multiple-blanks question
}
