% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{answer_none}
\alias{answer_none}
\title{A text-only question with no answers}
\usage{
answer_none()
}
\description{
A text-only question with no answers
}
\seealso{
Other answer-functions: 
\code{\link{answer_essay}()},
\code{\link{answer_matching}()},
\code{\link{answer_mcq}()},
\code{\link{answer_multiple}()},
\code{\link{answer_num}()},
\code{\link{answer_text}()},
\code{\link{answer_true_false}()},
\code{\link{answer_upload_file}()}
}
\concept{answer-functions}
