% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{delete_all_quizzes}
\alias{delete_all_quizzes}
\title{Delete all quizzes}
\usage{
delete_all_quizzes(
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\description{
Delete all quizzes
}
