% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer.R
\name{dropdown}
\alias{dropdown}
\title{Create a dropdown answer for a fill-in-multiple-blanks question}
\usage{
dropdown(correct, choices, id = NULL)
}
\arguments{
\item{correct}{Character vector of correct answer(s). Should be one element for single-answer multiple choice questions and can be multiple elements for multiple-answer multiple choice questions.}

\item{choices}{Character vector of answer choices.}

\item{blank_id}{The ID of the blank this dropdown corresponds to.}
}
\description{
Create a dropdown answer for a fill-in-multiple-blanks question
}
