% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{fill_in_the_blank}
\alias{fill_in_the_blank}
\title{Create a fill-in-the-blank answer for a fill-in-multiple-blanks question}
\usage{
fill_in_the_blank(text, id = NULL)
}
\arguments{
\item{text}{The correct answer text for this blank.}

\item{id}{The ID of the blank this answer corresponds to.}
}
\description{
Create a fill-in-the-blank answer for a fill-in-multiple-blanks question
}
