% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{list_folder}
\alias{list_folder}
\title{List the folder in a course}
\usage{
list_folder(
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\description{
List the folder in a course
}
