% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{list_submissions}
\alias{list_submissions}
\title{List quiz submissions}
\usage{
list_submissions(
  quiz_id,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN"),
  n = count_submissions(quiz_id, course_id, url, token),
  tz = Sys.timezone()
)
}
\arguments{
\item{quiz_id}{The ID of the quiz to retrieve statistics for.}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}

\item{n}{The maximum number of quiz submissions to retrieve.}

\item{tz}{The timezone to use for the started_at and finished_at columns. Defaults to the system timezone.}
}
\description{
List quiz submissions
}
