% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{submission_info}
\alias{submission_info}
\title{Get quiz submission details}
\usage{
submission_info(
  submission_id,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{submission_id}{The ID of the quiz submission to retrieve details for.}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\description{
Get quiz submission details
}
