% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission.R
\name{submission_overview}
\alias{submission_overview}
\title{Get quiz submission overview}
\usage{
submission_overview(
  submission_id,
  course_id = Sys.getenv("CANVASQUIZ_COURSE_ID"),
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{submission_id}{The ID of the quiz submission to retrieve details for.}

\item{course_id}{The course id. Defaults to the value of the \code{CANVASQUIZ_COURSE_ID} environment variable.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\value{
A data frame with the name of the user, quiz title, score, attempt number, quiz id, and submission id.
}
\description{
Get quiz submission overview
}
\seealso{
Other submissions: 
\code{\link{list_attempted_questions}()},
\code{\link{list_submissions}()},
\code{\link{submission_info}()},
\code{\link{submission_questions}()}
}
\concept{submissions}
