% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission.R
\name{submission_questions}
\alias{submission_questions}
\title{List attempted questions for a quiz}
\usage{
submission_questions(
  submission_id,
  url = Sys.getenv("CANVASQUIZ_URL"),
  token = Sys.getenv("CANVASQUIZ_TOKEN")
)
}
\arguments{
\item{submission_id}{The ID of the quiz submission to retrieve details for.}

\item{url}{The canvas url. Defaults to the value of the \code{CANVASQUIZ_URL} environment variable.}

\item{token}{The canvas token. Defaults to the value of the \code{CANVASQUIZ_TOKEN} environment variable.}
}
\description{
List attempted questions for a quiz
}
\seealso{
Other submissions: 
\code{\link{list_attempted_questions}()},
\code{\link{list_submissions}()},
\code{\link{submission_info}()},
\code{\link{submission_overview}()}
}
\concept{submissions}
