% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tag_img}
\alias{tag_img}
\alias{upload_tag_img}
\title{Generate an HTML image tag for a file in a Canvas course}
\usage{
tag_img(
  file_id,
  width = "100\%",
  class = "",
  course_id = Sys.getenv("CANVAS_COURSE_ID")
)

upload_tag_img(
  file_path,
  width = "100\%",
  class = "",
  folder_id = NULL,
  course_id = Sys.getenv("CANVAS_COURSE_ID"),
  url = Sys.getenv("CANVAS_URL"),
  token = Sys.getenv("CANVAS_TOKEN")
)
}
\arguments{
\item{file_id}{The file ID of the image in Canvas.}

\item{width}{The width of the image (default is "100\%").}

\item{class}{Optional CSS class to apply to the \verb{<img>} tag.}

\item{course_id}{The ID of the Canvas course where the file is stored (default is taken from the environment variable \code{CANVAS_COURSE_ID}).}
}
\value{
A character string containing the HTML \verb{<img>} tag for the specified image file.
}
\description{
This function creates an HTML \verb{<img>} tag that can be used to display an image file stored in a Canvas course. The image will be displayed at 50\% width.
}
